refactored Component files
